feat(components): add `post-side-navigation` styles
