+++ .changeset/nice-games-crash.md
@@ -1,0 +1,6 @@
+---
+'@swisspost/design-system-components': minor
+'@swisspost/design-system-documentation': minor
+---
+
+Added the `post-side-navigation` component.
