+++ apps/documentation/src/stories/components/side-navigation/side-navigation.styles.scss
@@ -1,0 +1,3 @@
+.docs-story > :first-child {
+  overflow: hidden;
+}
