--- apps/documentation/src/stories/raw-components/side-navigation/side-navigation.styles.scss
@@ -1,4 +1,0 @@
-.side-nav-story-wrapper {
-    overflow: auto;
-    height: 700px;
-}
