+++ packages/components/cypress/e2e/side-navigation.cy.ts
@@ -39,6 +39,26 @@ describe('post-side-navigation', () => {
       cy.get('@side-navigation').then(([el]) => el.toggle());
       cy.get('@side-navigation').shadow().find('dialog').should('not.exist');
     });
+
+    it('should not have the small class by default', () => {
+      cy.get('@side-navigation').should('not.have.class', 'post-side-navigation-small');
+    });
+  });
+
+  describe('isSmall prop', () => {
+    beforeEach(() => {
+      cy.viewport(1280, 800);
+      cy.getComponents(
+        SIDENAVIGATION_ID,
+        'small',
+        'post-side-navigation',
+        'post-side-navigation-trigger',
+      );
+    });
+
+    it('should add the small class when isSmall is true', () => {
+      cy.get('@side-navigation').should('have.class', 'post-side-navigation-small');
+    });
   });
 
   describe('mobile', () => {
